cmd: gdb: Add `disconnect` after dumping the image for verification.
By disconnecting, the target will not be restarted if it was halted.
This is needed if connecting to a qemu instances started with `-S`, as
humility would connect, dump the image, and detach, restarting the
target.  Then reconnect with gdb for the user, but this would not be at
the first instruction.  Now on the second connection, the VM will still
be halted at the first instruction.