One-shot tasks (#13)
* Print "STOPPED" instead of "not started" for SchedState::Stopped

This change updates explain_sched_state to print "STOPPED" instead of "not
started" to print out the state name rather than indicating that the task didn't
start. With the changes for supporting one-shot tasks (i.e. tasks that call exit
to move to stopped state), we might see tasks getting into a stopped state after
they are done with their work as well.