humility-core: core: GDBCore: parse feature xml provided by gdbstub.
Qemu may use arbitrary values for the csr numbers, so we cannot code
them into humility.  Instead we fetch the `target.xml` from the gdbstub.
This xml contains the mapping of register names to numbers. If we
receive this mapping, humility will use it to read registers, otherwise
it will fall back to the default GDB ids.  If the mapping is recieved,
we will only read registers included in it.